Add committee store & pubsub indexing

+++ store/migrations/2023-07-14-181200_epoch/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS "epoch" (
     "index" NUMERIC(20,0) PRIMARY KEY NOT NULL,
-    "active_validators" BIGINT NOT NULL,
-    "total_validators" BIGINT NOT NULL
+    "active_validators" NUMERIC(20,0) NOT NULL,
+    "total_validators" NUMERIC(20,0) NOT NULL
 );
