+++ store/migrations/2023-07-14-181205_validator/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS "validator" (
-    "index" BIGINT PRIMARY KEY NOT NULL,
+    "index" NUMERIC(20,0) PRIMARY KEY NOT NULL,
     "pubkey" VARCHAR NOT NULL,
     "activation_epoch" NUMERIC(20,0) NOT NULL,
     "exit_epoch" NUMERIC(20,0) NOT NULL
